close bin/1648 libmd not 64bit safe.
if something fails to compile now, you need to add #include <sys/types.h>

Partially Submitted by:	Jason Thorpe <thorpej@nas.nasa.gov>


git-svn-id: svn+ssh://svn.freebsd.org/base/head@19099 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ lib/libmd/md4.h
@@ -27,8 +27,8 @@
 #define _MD4_H_
 /* MD4 context. */
 typedef struct MD4Context {
-  unsigned long state[4];	/* state (ABCD) */
-  unsigned long count[2];	/* number of bits, modulo 2^64 (lsb first) */
+  u_int32_t state[4];	/* state (ABCD) */
+  u_int32_t count[2];	/* number of bits, modulo 2^64 (lsb first) */
   unsigned char buffer[64];	/* input buffer */
 } MD4_CTX;
